back-end/requirements.txt
Django>=4.2
djangorestframework>=3.14
Pillow>=10.0
numpy>=1.24
opencv-python>=4.7
face-recognition>=1.3
dlib>=19.24
opencv-python-headless
opencv-contrib-python-headless
numpy
psycopg2 